requirements.txt
# Multi-Agent Software Engineering System
# Dependencies pinned to exact versions

# --- LangChain / LangGraph ---
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# --- LLM Provider ---
google-generativeai>=0.8.0

# --- Observability ---
langsmith>=0.1.0

# --- GitHub API ---
PyGithub>=2.4.0

# --- Docker SDK ---
docker>=7.1.0

# --- Data Validation ---
pydantic>=2.9.0

# --- Environment Variables ---
python-dotenv>=1.0.0

# --- Testing ---
pytest>=8.3.0

# --- API Server ---
fastapi>=0.115.0
uvicorn>=0.30.0

# --- Code Quality ---
black>=24.0.0
ruff>=0.6.0
